доработки по предыдущему заданию
