--- src/components/constants.js
@@ -1,23 +1,0 @@
-export const DESCRIPTIONS = [`Изучить теорию`, `Сделать домашку`, `Пройти интенсив на соточку`];
-export const TAGS = [`homework`, `theory`, `practice`, `intensive`, `keks`];
-export const TIME_WEEK = 7 * 24 * 60 * 60 * 1000;
-export const TIME_DAY = 24 * 60 * 60 * 1000;
-export const TASKS_AMOUNT = 16;
-
-export const TaskDay = {
-  MO: `mo`,
-  TU: `tu`,
-  WE: `we`,
-  TH: `th`,
-  FR: `fr`,
-  SA: `sa`,
-  SU: `su`
-};
-
-export const TaskColor = {
-  BLACK: `black`,
-  YELLOW: `yellow`,
-  BLUE: `blue`,
-  GREEN: `green`,
-  PINK: `pink`
-};
