+++ src/components/index.js
@@ -2,7 +2,6 @@ import {getMenuTemplate} from './site-menu';
 import {getSearchTemplate} from './search';
 import {getFilters} from './filters';
 import {getBoardTasksTemplate} from './board-task';
-import {tasksRendered} from './data';
 import {renderTasks, noTasksTemplate} from './task';
 
 export {
@@ -11,6 +10,5 @@ export {
   getFilters,
   getBoardTasksTemplate,
   renderTasks,
-  tasksRendered,
   noTasksTemplate
 };
